Changed Grafana configuration to enable more default permissions for automatically created user accounts.
Authentication and access restriction is performed by Apache. The idea
of this whole setup is to keep things as simple as possible. So if you
enable access for some coworker, account will be automatically created
in Grafana with quite powerfull initial permissions, so there is no
action required by administrator.